+++ src/main/java/com/sql/sqlextra/entity/EmailSent.java
@@ -17,7 +17,7 @@ public class EmailSent {
     private Long id;
 
     @Column(name = "id_account", nullable = false)
-    private Integer idAccount;
+    private Long idAccount;
 
     @Column(name = "sent_date")
     private Integer sentDate;
